requirements.txt
streamlit
pandas
pydantic
requests
beautifulsoup4
googlesearch-python
